RavenDB-22456 - Create Client API
