+++ src/Raven.Server/Documents/Handlers/Processors/Revisions/RevisionsHandlerProcessorForRevertRevisionsForDocument.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
+﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using JetBrains.Annotations;
 using Raven.Server.ServerWide;
